Add account deletion functionality and refactor auth actions
- all the auth actions are now moved under auth folder
- users can delete their account and it will delete their data along with it (as long as have on cascade delete setup on other tables)
- created components folder

Note: In order to delete a user from the auth table we need the service role key in supabase so if you are testing this please replace the public key with service role key in the env file
